requirements.txt
ollama
gitpython>=3.1.40
python-dotenv>=1.0.0
typing-extensions>=4.0.0
pathlib>=1.0.1
argparse>=1.4.0